Move dump related code to its own file
