+++ src/apps/lwaftr/lwaftr.lua
@@ -2,6 +2,7 @@ module(..., package.seeall)
 
 local bt = require("apps.lwaftr.binding_table")
 local constants = require("apps.lwaftr.constants")
+local dump = require('apps.lwaftr.dump')
 local fragmentv4 = require("apps.lwaftr.fragmentv4")
 local fragmentv6 = require("apps.lwaftr.fragmentv6")
 local icmp = require("apps.lwaftr.icmp")
@@ -27,14 +28,10 @@ local C = ffi.C
 local cast, fstring = ffi.cast, ffi.string
 local receive, transmit = link.receive, link.transmit
 local rd16, rd32, get_ihl_from_offset = lwutil.rd16, lwutil.rd32, lwutil.get_ihl_from_offset
-local set, keys, write_to_file = lwutil.set, lwutil.keys, lwutil.write_to_file
-local ipv4number_to_str = lwutil.ipv4number_to_str
+local keys = lwutil.keys
 local write_eth_header, write_ipv6_header = lwheader.write_eth_header, lwheader.write_ipv6_header 
 
 local debug = false
-
-local CONF_FILE_DUMP = "/tmp/lwaftr-%s.conf"
-local BINDING_TABLE_FILE_DUMP = "/tmp/binding-%s.table"
 
 local function compute_binding_table_by_ipv4(binding_table)
    local ret = {}
@@ -126,69 +123,6 @@ local function reload_binding_table(lwstate)
    lwstate.binding_table_by_ipv4 = compute_binding_table_by_ipv4(lwstate.binding_table)
 end
 
-local function dump_configuration(lwstate)
-   print("Dump configuration")
-   local result = {}
-   local etharr = set('aftr_mac_b4_side',  'aftr_mac_inet_side', 'b4_mac',  'inet_mac')
-   local ipv4arr = set('aftr_ipv4_ip')
-   local ipv6arr = set('aftr_ipv6_ip')
-   local val
-   for _, k in ipairs(lwstate.conf_keys) do
-      local v = lwstate[k]
-      if etharr[k] then
-         val = ("ethernet:pton('%s')"):format(ethernet:ntop(v))
-      elseif ipv4arr[k] then
-         val = ("ipv4:pton('%s')"):format(ipv4:ntop(v))
-      elseif ipv6arr[k] then
-         val = ("ipv6:pton('%s')"):format(ipv6:ntop(v))
-      elseif type(v) == "bool" then
-         val = v and "true" or "false"
-      elseif k == "binding_table" then
-         val = "bt.get_binding_table()"
-      else
-         val = lwstate[k]
-      end
-      table.insert(result, ("%s = %s"):format(k, val))
-   end
-   local filename = (CONF_FILE_DUMP):format(os.date("%Y-%m-%d-%H:%M:%S"))
-   local content = table.concat(result, ",\n")
-   write_to_file(filename, content)
-   print(("Configuration written to %s"):format(filename))
-end
-
-local function dump_binding_table(lwstate)
-   print("Dump binding table")
-   local content = {}
-   local function write(str)
-      table.insert(content, str)
-   end
-   local function dump()
-      return table.concat(content, "\n")
-   end
-   local function format_entry(entry)
-      local v6, v4, port_start, port_end, br_v6 = entry[1], entry[2], entry[3], entry[4], entry[5]
-      local result = {}
-      table.insert(result, ("'%s'"):format(ipv6:ntop(v6)))
-      table.insert(result, ("'%s'"):format(ipv4number_to_str(v4)))
-      table.insert(result, port_start)
-      table.insert(result, port_end)
-      if br_v6 then
-         table.insert(result, ("'%s'"):format(ipv6:ntop(br_v6)))
-      end
-      return table.concat(result, ",")
-   end
-   -- Write entries to content
-   write("{")
-   for _, entry in ipairs(lwstate.binding_table) do
-      write(("\t{%s},"):format(format_entry(entry)))
-   end
-   write("}")
-   -- Dump content to file
-   local filename = (BINDING_TABLE_FILE_DUMP):format(os.date("%Y-%m-%d-%H:%M:%S"))
-   write_to_file(filename, dump())
-   print(("Binding table written to %s"):format(filename))
-end
-
 LwAftr = {}
 
 function LwAftr:new(conf)
@@ -218,8 +152,8 @@ function LwAftr:new(conf)
    transmit_icmpv6_with_rate_limit = init_transmit_icmpv6_with_rate_limit(o)
    on_signal("hup", function() reload_binding_table(o) end)
    on_signal("usr1", function()
-      dump_configuration(o)
-      dump_binding_table(o)
+      dump.dump_configuration(o)
+      dump.dump_binding_table(o)
    end)
    o.conf_keys = keys(conf)
    if debug then lwdebug.pp(conf) end
